requirements.txt
# Core scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Agent / Claude API
anthropic>=0.25.0

# Data validation (used in agent/models.py)
pydantic>=2.0.0

# Load .env files for API keys
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
